+++ aui.views/src/AUI/Platform/AWindows.cpp
@@ -133,7 +133,7 @@ void AWindow::onFocusAcquired() {
     AViewContainer::onFocusAcquired();
 }
 
-void AWindow::onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event) {
+void AWindow::onPointerMove(glm::vec2 pos, const APointerMoveEvent& event) {
     ABaseWindow::onPointerMove(pos, event);
 }
