+++ aui.views/src/AUI/Platform/android/AndroidJniInterface.cpp
@@ -43,7 +43,7 @@ Java_com_github_aui_android_AuiView_handleResize(JNIEnv *env, jclass clazz, jint
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_github_aui_android_AuiView_handleLongPress(JNIEnv *env, jclass clazz, jint x, jint y) {
+Java_com_github_aui_android_AuiView_handleLongPress(JNIEnv *env, jclass clazz, jfloat x, jfloat y) {
     runOnGLThread([=] {
         AUI_NULLSAFE(AWindow::current())->onGesture({x, y}, ALongPressEvent{});
     });
@@ -51,16 +51,16 @@ Java_com_github_aui_android_AuiView_handleLongPress(JNIEnv *env, jclass clazz, j
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_github_aui_android_AuiView_handlePointerButtonDown(JNIEnv *env, jclass clazz, jint x,
-                                                          jint y, jint pointerId) {
+Java_com_github_aui_android_AuiView_handlePointerButtonDown(JNIEnv *env, jclass clazz, jfloat x,
+                                                          jfloat y, jint pointerId) {
     runOnGLThread([=] {
         AUI_NULLSAFE(AWindow::current())->onPointerPressed({{x, y}, APointerIndex::finger(pointerId)});
     });
 }
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_github_aui_android_AuiView_handlePointerButtonUp(JNIEnv *env, jclass clazz, jint x,
-                                                        jint y, jint pointerId) {
+Java_com_github_aui_android_AuiView_handlePointerButtonUp(JNIEnv *env, jclass clazz, jfloat x,
+                                                        jfloat y, jint pointerId) {
     runOnGLThread([=] {
         AUI_NULLSAFE(AWindow::current())->onPointerReleased({{x, y}, APointerIndex::finger(pointerId)});
     });
@@ -68,8 +68,8 @@ Java_com_github_aui_android_AuiView_handlePointerButtonUp(JNIEnv *env, jclass cl
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_github_aui_android_AuiView_handlePointerMove(JNIEnv *env, jclass clazz, jint x,
-                                                    jint y, jint pointerId) {
+Java_com_github_aui_android_AuiView_handlePointerMove(JNIEnv *env, jclass clazz, jfloat x,
+                                                    jfloat y, jint pointerId) {
     // android sends move events for all fingers even if just one finger moved.
     static std::unordered_map<jint /* pointerIndex */, glm::ivec2> prevValues;
     glm::ivec2 currentValue{x, y};
