+++ aui.views/src/AUI/View/AAbstractTypeableView.cpp
@@ -351,7 +351,7 @@ void AAbstractTypeableView::onPointerPressed(const APointerPressedEvent& event)
     updateCursorBlinking();
 }
 
-void AAbstractTypeableView::onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event)
+void AAbstractTypeableView::onPointerMove(glm::vec2 pos, const APointerMoveEvent& event)
 {
     AView::onPointerMove(pos, event);
     ACursorSelectable::handleMouseMove(pos);
