+++ aui.views/src/AUI/View/ADragNDropView.h
@@ -23,7 +23,7 @@ class API_AUI_VIEWS ADragNDropView: public AViewContainer {
 public:
     ADragNDropView();
 
-    void onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event) override;
+    void onPointerMove(glm::vec2 pos, const APointerMoveEvent& event) override;
 
 private:
     bool mDnDInProgress = false;
