+++ aui.views/src/AUI/View/ARulerArea.h
@@ -42,7 +42,7 @@ protected:
 public:
     ARulerArea(const _<AView>& wrappedView);
     void setSize(glm::ivec2 size) override;
-    void onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event) override;
+    void onPointerMove(glm::vec2 pos, const APointerMoveEvent& event) override;
     void render() override;
 };
