+++ aui.views/src/AUI/View/ATreeView.cpp
@@ -165,7 +165,7 @@ public:
         mTreeView->handleMouseDoubleClicked(this);
     }
 
-    void onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event) override {
+    void onPointerMove(glm::vec2 pos, const APointerMoveEvent& event) override {
         AViewContainer::onPointerMove(pos, event);
         mTreeView->handleMouseMove(this);
     }
