+++ aui.views/src/AUI/View/AView.cpp
@@ -335,7 +335,7 @@ void AView::onMouseEnter()
 }
 
 
-void AView::onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event)
+void AView::onPointerMove(glm::vec2 pos, const APointerMoveEvent& event)
 {
     AWindow::current()->setCursor(mCursor);
 }
