+++ aui.views/src/AUI/View/AViewContainer.h
@@ -96,7 +96,7 @@ public:
 
     void onMouseEnter() override;
 
-    void onPointerMove(glm::ivec2 pos, const APointerMoveEvent& event) override;
+    void onPointerMove(glm::vec2 pos, const APointerMoveEvent& event) override;
 
     void onMouseLeave() override;
